AMBARI-25379 Upgrade AMS Grafana version to 6.3.5
The commit of BUG-114430 has been cherry-picked and the datasource.js was backported.
This way Grafana is first upgraded to 5.2.4 as it was done already on the 2.8 and 2.7.100
branches. Probably it is going to be easier to upgrade from 5.2.4 to 6.3.5 than go from 2.x
in one step.

Change-Id: I720f5fc5e6fbd82bf4f68e0631a11fc4967a27a0
(cherry picked from commit c12203b88dfb10e725a6c69a2f792abf20b2669f)